+++ system/src/Grav/Common/Grav.php
@@ -8,6 +8,7 @@ use Grav\Common\Service\ConfigServiceProvider;
 use Grav\Common\Service\ErrorServiceProvider;
 use Grav\Common\Service\LoggerServiceProvider;
 use Grav\Common\Service\StreamsServiceProvider;
+use Grav\Common\Twig\Twig;
 use RocketTheme\Toolbox\DI\Container;
 use RocketTheme\Toolbox\Event\Event;
 use RocketTheme\Toolbox\Event\EventDispatcher;
@@ -219,10 +220,7 @@ class Grav extends Container
         $this['pages']->init();
         $this->fireEvent('onPagesInitialized');
         $debugger->stopTimer('pages');
-
-        $debugger->startTimer('pageinit', 'Page Initialized');
         $this->fireEvent('onPageInitialized');
-        $debugger->stopTimer('pageinit');
 
         $debugger->addAssets();
 
@@ -350,6 +348,11 @@ class Grav extends Container
         if (isset($this['page']->header()->http_response_code)) {
             http_response_code($this['page']->header()->http_response_code);
         }
+
+        // Vary: Accept-Encoding
+        if ($this['config']->get('system.pages.vary_accept_encoding', false)) {
+            header('Vary: Accept-Encoding');
+        }
     }
 
     /**
@@ -438,7 +441,17 @@ class Grav extends Container
             }
 
             // unsupported media type, try to download it...
-            $extension = $uri->extension() ?: isset($path_parts['extension']) ? $path_parts['extension'] : null;
+            $uri_extension = $uri->extension();
+            if ($uri_extension) {
+                $extension = $uri_extension;
+            } else {
+                if (isset($path_parts['extension'])) {
+                    $extension = $path_parts['extension'];
+                } else {
+                    $extension = null;
+                }
+            }
+
             if ($extension) {
                 $download = true;
                 if (in_array(ltrim($extension, '.'), $this['config']->get('system.media.unsupported_inline_types'))) {